Move everything except JdbcSourceConnector and the util pkg to under a 'source' subpkg
In anticipation of another 'sink' subpkg
