+++ src/main/java/io/confluent/connect/jdbc/source/JdbcUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  **/
 
-package io.confluent.connect.jdbc;
+package io.confluent.connect.jdbc.source;
 
 import org.apache.kafka.connect.errors.ConnectException;
 import org.slf4j.Logger;
