+++ src/test/java/io/confluent/connect/jdbc/JdbcSourceConnectorTest.java
@@ -35,6 +35,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import io.confluent.connect.jdbc.source.EmbeddedDerby;
+import io.confluent.connect.jdbc.source.JdbcSourceConnectorConfig;
+import io.confluent.connect.jdbc.source.JdbcSourceTask;
+import io.confluent.connect.jdbc.source.JdbcSourceTaskConfig;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
