+++ src/test/java/io/confluent/connect/jdbc/source/JdbcUtilsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  **/
 
-package io.confluent.connect.jdbc;
+package io.confluent.connect.jdbc.source;
 
 import org.junit.After;
 import org.junit.Before;
