+++ src/test/java/io/confluent/connect/jdbc/source/TableMonitorThreadTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  **/
 
-package io.confluent.connect.jdbc;
+package io.confluent.connect.jdbc.source;
 
 import org.apache.kafka.connect.connector.ConnectorContext;
 import org.easymock.EasyMock;
