+++ src/test/java/io/confluent/connect/jdbc/source/TimestampIncrementingOffsetTest.java
@@ -14,12 +14,14 @@
  * limitations under the License.
  **/
 
-package io.confluent.connect.jdbc;
+package io.confluent.connect.jdbc.source;
 
 import org.junit.Before;
 import org.junit.Test;
 
 import java.sql.Timestamp;
+
+import io.confluent.connect.jdbc.source.TimestampIncrementingOffset;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
